# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API EliteRT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rumah": {
            "get": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Menampilkan semua data rumah",
                "operationId": "ba129de8b8f3bfed0ff4ce5b3a170e79",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data rumah"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Menambahkan rumah baru",
                "operationId": "6adf8ec6d8a98f552f2939b30e250f60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_rumah",
                                    "no_rumah",
                                    "alamat_rumah",
                                    "status_rumah"
                                ],
                                "properties": {
                                    "nama_rumah": {
                                        "type": "string"
                                    },
                                    "no_rumah": {
                                        "type": "string"
                                    },
                                    "alamat_rumah": {
                                        "type": "string"
                                    },
                                    "status_rumah": {
                                        "type": "string",
                                        "enum": [
                                            "Tetap",
                                            "Kontrak",
                                            "Tidak dihuni"
                                        ]
                                    },
                                    "kondisi_rumah": {
                                        "type": "string",
                                        "enum": [
                                            "Dihuni",
                                            "Tidak Dihuni"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rumah berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/rumah/{id}": {
            "get": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Menampilkan detail rumah berdasarkan ID",
                "operationId": "859ad098a1ca3f9fbc9afa9f84546397",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail rumah"
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Memperbarui data rumah",
                "operationId": "f47610ecaa510a425deef73a72982924",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_rumah",
                                    "no_rumah",
                                    "alamat_rumah",
                                    "status_rumah"
                                ],
                                "properties": {
                                    "nama_rumah": {
                                        "type": "string"
                                    },
                                    "no_rumah": {
                                        "type": "string"
                                    },
                                    "alamat_rumah": {
                                        "type": "string"
                                    },
                                    "status_rumah": {
                                        "type": "string",
                                        "enum": [
                                            "Tetap",
                                            "Kontrak",
                                            "Tidak dihuni"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rumah berhasil diperbarui"
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Menghapus data rumah",
                "operationId": "2322ebc25b0cdd54ee80b313ff3881ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rumah berhasil dihapus"
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan"
                    }
                }
            }
        }
    },
...
}